Misc
Misc

Refactor base feature widget

M2

Worked

Misc

Allow changing default resolution and normalization in hic
