+++ packages/core/BaseFeatureWidget/BaseFeatureDetail/index.tsx
@@ -1,132 +1,17 @@
 import React from 'react'
-import { ErrorBoundary } from '@jbrowse/core/ui/ErrorBoundary'
-import {
-  Accordion,
-  AccordionDetails,
-  AccordionSummary,
-  Divider,
-  Typography,
-} from '@mui/material'
-import { makeStyles } from 'tss-react/mui'
 import { observer } from 'mobx-react'
-import { IAnyStateTreeNode } from 'mobx-state-tree'
 
-// icons
-import ExpandMore from '@mui/icons-material/ExpandMore'
+// utils
+import { ErrorMessage } from '../../ui'
+import { BaseCardProps, BaseProps } from '../types'
+import { isEmpty } from './util'
+import { replaceUndefinedWithNull } from '../util'
 
 // locals
-import {
-  getEnv,
-  getSession,
-  assembleLocString,
-  toLocale,
-  SimpleFeatureSerialized,
-} from '../../util'
-import { ErrorMessage } from '../../ui'
-import SequenceFeatureDetails from '../SequenceFeatureDetails'
-import { BaseCardProps, BaseProps } from '../types'
-import SimpleField from './SimpleField'
 import Attributes from './Attributes'
-import { generateTitle, isEmpty } from './util'
-
-// coreDetails are omitted in some circumstances
-const coreDetails = [
-  'name',
-  'start',
-  'end',
-  'strand',
-  'refName',
-  'description',
-  'type',
-]
-
-const useStyles = makeStyles()(theme => ({
-  expansionPanelDetails: {
-    display: 'block',
-    padding: theme.spacing(1),
-  },
-  icon: {
-    color: theme.palette.tertiary.contrastText || '#fff',
-  },
-}))
-
-export function BaseCard({
-  children,
-  title,
-  defaultExpanded = true,
-}: BaseCardProps) {
-  const { classes } = useStyles()
-  return (
-    <Accordion defaultExpanded={defaultExpanded}>
-      <AccordionSummary expandIcon={<ExpandMore className={classes.icon} />}>
-        <Typography variant="button">{title}</Typography>
-      </AccordionSummary>
-      <AccordionDetails className={classes.expansionPanelDetails}>
-        {children}
-      </AccordionDetails>
-    </Accordion>
-  )
-}
-
-function Position(props: BaseProps) {
-  const { feature } = props
-  const strand = feature.strand as number
-  const strandMap: Record<string, string> = {
-    '-1': '-',
-    '0': '',
-    '1': '+',
-  }
-  const str = strandMap[strand] ? `(${strandMap[strand]})` : ''
-  const loc = assembleLocString(feature)
-  return <>{`${loc} ${str}`}</>
-}
-
-function CoreDetails(props: BaseProps) {
-  const { feature } = props
-  const obj = feature as SimpleFeatureSerialized & {
-    start: number
-    end: number
-    assemblyName?: string
-    strand: number
-    refName: string
-    __jbrowsefmt: {
-      start?: number
-      assemblyName?: string
-      end?: number
-      refName?: string
-      name?: string
-    }
-  }
-
-  const formattedFeat = { ...obj, ...obj.__jbrowsefmt }
-  const { start, end } = formattedFeat
-
-  const displayedDetails: Record<string, any> = {
-    ...formattedFeat,
-    length: toLocale(end - start),
-  }
-
-  const coreRenderedDetails = {
-    description: 'Description',
-    name: 'Name',
-    length: 'Length',
-    type: 'Type',
-  }
-  return (
-    <>
-      <SimpleField
-        name="Position"
-        value={<Position {...props} feature={formattedFeat} />}
-      />
-      {Object.entries(coreRenderedDetails)
-        .map(([key, name]) => [name, displayedDetails[key]])
-        .filter(([, value]) => value != null)
-        .map(([name, value]) => (
-          <SimpleField key={name} name={name} value={value} />
-        ))}
-    </>
-  )
-}
+import BaseCard from './BaseCard'
+import CoreDetails from './CoreDetails'
+import FeatureDetails from './FeatureDetails'
 
 export const BaseCoreDetails = (props: BaseProps) => {
   const { title = 'Primary data' } = props
@@ -153,113 +38,23 @@ export interface BaseInputProps extends BaseCardProps {
   formatter?: (val: unknown, key: string) => React.ReactNode
 }
 
-interface PanelDescriptor {
-  name: string
-  Component: React.FC<any>
-}
-
-export function FeatureDetails(props: {
-  model: IAnyStateTreeNode
-  feature: SimpleFeatureSerialized
-  depth?: number
-  omit?: string[]
-  descriptions?: Record<string, React.ReactNode>
-  formatter?: (val: unknown, key: string) => React.ReactNode
-}) {
-  const { omit = [], model, feature, depth = 0 } = props
-  const { maxDepth } = model
-  const { mate, name = '', id = '', type = '', subfeatures, uniqueId } = feature
-  const pm = getEnv(model).pluginManager
-  const session = getSession(model)
-
-  const ExtraPanel = pm.evaluateExtensionPoint('Core-extraFeaturePanel', null, {
-    session,
-    feature,
-    model,
-  }) as PanelDescriptor | undefined
-  const m = mate as { start: number; end: number; refName: string } | undefined
-  return (
-    <BaseCard title={generateTitle(name, id, type)}>
-      <Typography>Core details</Typography>
-      <CoreDetails {...props} />
-      {m ? (
-        <>
-          <Divider />
-          <Typography>Mate details</Typography>
-          <CoreDetails
-            {...props}
-            feature={{
-              ...m,
-              start: m.start,
-              end: m.end,
-              refName: m.refName,
-              uniqueId: `${uniqueId}-mate`,
-            }}
-          />
-        </>
-      ) : null}
-
-      <Divider />
-      <Typography>Attributes</Typography>
-      <Attributes
-        attributes={feature}
-        {...props}
-        omit={omit}
-        omitSingleLevel={coreDetails}
-      />
-
-      <ErrorBoundary FallbackComponent={e => <ErrorMessage error={e.error} />}>
-        <SequenceFeatureDetails {...props} />
-      </ErrorBoundary>
-
-      {ExtraPanel ? (
-        <>
-          <Divider />
-          <BaseCard title={ExtraPanel.name}>
-            <ExtraPanel.Component {...props} />
-          </BaseCard>
-        </>
-      ) : null}
-
-      {depth < maxDepth && subfeatures?.length ? (
-        <BaseCard title="Subfeatures" defaultExpanded={depth < 1}>
-          {subfeatures.map((sub, idx) => (
-            <FeatureDetails
-              key={JSON.stringify(sub)}
-              feature={{
-                ...sub,
-                uniqueId: `${uniqueId}_${idx}`,
-              }}
-              model={model}
-              depth={depth + 1}
-            />
-          ))}
-        </BaseCard>
-      ) : null}
-    </BaseCard>
-  )
-}
-
 const BaseFeatureDetail = observer(function ({ model }: BaseInputProps) {
   const { error, featureData } = model
 
   if (error) {
     return <ErrorMessage error={error} />
-  }
-  if (!featureData) {
+  } else if (!featureData) {
     return null
+  } else {
+    // replacing undefined with null helps with allowing fields to be hidden,
+    // setting null is not allowed by jexl so we set it to undefined to hide.
+    // see config guide. this replacement happens both here and when
+    // snapshotting the featureData
+    const featureData2 = replaceUndefinedWithNull(featureData)
+    return isEmpty(featureData2) ? null : (
+      <FeatureDetails model={model} feature={featureData2} />
+    )
   }
-
-  // replacing undefined with null helps with allowing fields to be hidden,
-  // setting null is not allowed by jexl so we set it to undefined to hide. see
-  // config guide. this replacement happens both here and when snapshotting the
-  // featureData
-  const g = JSON.parse(
-    JSON.stringify(featureData, (_, v) => (v === undefined ? null : v)),
-  )
-  return isEmpty(g) ? null : <FeatureDetails model={model} feature={g} />
 })
 
 export default BaseFeatureDetail
-
-export { default as Attributes } from './Attributes'
