--- packages/core/ReExports/Attributes.tsx
@@ -1,3 +1,0 @@
-// convert named export to default for lazy react
-
-export { Attributes as default } from '@jbrowse/core/BaseFeatureWidget/BaseFeatureDetail'
