--- packages/core/ReExports/BaseCard.tsx
@@ -1,3 +1,0 @@
-// convert named export to default for lazy react
-
-export { BaseCard as default } from '@jbrowse/core/BaseFeatureWidget/BaseFeatureDetail'
