--- packages/core/ReExports/FeatureDetails.tsx
@@ -1,3 +1,0 @@
-// convert named export to default for lazy react
-
-export { FeatureDetails as default } from '@jbrowse/core/BaseFeatureWidget/BaseFeatureDetail'
