+++ packages/product-core/src/ui/AboutDialogContents.tsx
@@ -10,10 +10,10 @@ import {
   AnyConfigurationModel,
 } from '@jbrowse/core/configuration'
 import { getSession, getEnv } from '@jbrowse/core/util'
-import {
-  BaseCard,
-  Attributes,
-} from '@jbrowse/core/BaseFeatureWidget/BaseFeatureDetail'
+import Attributes from '@jbrowse/core/BaseFeatureWidget/BaseFeatureDetail/Attributes'
+import BaseCard from '@jbrowse/core/BaseFeatureWidget/BaseFeatureDetail/BaseCard'
+
+// locals
 import FileInfoPanel from './FileInfoPanel'
 import RefNameInfoDialog from './RefNameInfoDialog'
