+++ packages/product-core/src/ui/FileInfoPanel.tsx
@@ -4,10 +4,8 @@ import {
   AnyConfigurationModel,
 } from '@jbrowse/core/configuration'
 import { getSession } from '@jbrowse/core/util'
-import {
-  BaseCard,
-  Attributes,
-} from '@jbrowse/core/BaseFeatureWidget/BaseFeatureDetail'
+import Attributes from '@jbrowse/core/BaseFeatureWidget/BaseFeatureDetail/Attributes'
+import BaseCard from '@jbrowse/core/BaseFeatureWidget/BaseFeatureDetail/BaseCard'
 import { ErrorMessage, LoadingEllipses } from '@jbrowse/core/ui'
 
 type FileInfo = Record<string, unknown> | string
