+++ packages/text-indexing/src/types/gff3Adapter.ts
@@ -23,17 +23,11 @@ export async function* indexGff3({
   onUpdate: (progressBytes: number) => void
 }) {
   const { trackId } = config
-
-  let receivedBytes = 0
-  const { totalBytes, stream } = await getLocalOrRemoteStream(
-    inLocation,
-    outDir,
-  )
-  onStart(totalBytes)
-
-  stream.on('data', chunk => {
-    receivedBytes += chunk.length
-    onUpdate(receivedBytes)
+  const stream = await getLocalOrRemoteStream({
+    file: inLocation,
+    out: outDir,
+    onTotalBytes: onStart,
+    onBytesReceived: onUpdate,
   })
 
   const rl = readline.createInterface({
