+++ plugins/alignments/src/AlignmentsFeatureDetail/AlignmentsFeatureDetail.tsx
@@ -2,7 +2,7 @@ import React, { lazy } from 'react'
 import { Paper } from '@mui/material'
 import { observer } from 'mobx-react'
 import clone from 'clone'
-import { FeatureDetails } from '@jbrowse/core/BaseFeatureWidget/BaseFeatureDetail'
+import FeatureDetails from '@jbrowse/core/BaseFeatureWidget/BaseFeatureDetail/FeatureDetails'
 
 // locals
 import { getTag } from './util'
