+++ plugins/linear-comparative-view/src/LGVSyntenyDisplay/model.ts
@@ -2,8 +2,15 @@ import { lazy } from 'react'
 import {
   ConfigurationReference,
   AnyConfigurationSchemaType,
+  getConf,
 } from '@jbrowse/core/configuration'
-import { getSession } from '@jbrowse/core/util'
+import {
+  Feature,
+  getContainingTrack,
+  getContainingView,
+  getSession,
+  isSessionModelWithWidgets,
+} from '@jbrowse/core/util'
 import { SharedLinearPileupDisplayMixin } from '@jbrowse/plugin-alignments'
 import { types } from 'mobx-state-tree'
 
@@ -48,12 +55,13 @@ function stateModelFactory(schema: AnyConfigurationSchemaType) {
             ...(feature
               ? [
                   {
-                    label: 'Open synteny view for this position',
+                    label: 'Launch synteny view for this position',
                     onClick: () => {
                       getSession(self).queueDialog(handleClose => [
                         LaunchSyntenyViewDialog,
                         {
                           model: self,
+                          trackId: getConf(getContainingTrack(self), 'trackId'),
                           handleClose,
                           feature,
                         },
@@ -87,6 +95,30 @@ function stateModelFactory(schema: AnyConfigurationSchemaType) {
       }
     })
     .actions(self => ({
+      /**
+       * #action
+       */
+      selectFeature(feature: Feature) {
+        const session = getSession(self)
+        if (isSessionModelWithWidgets(session)) {
+          const r2 = getContainingView(self)
+          let r3 = r2
+          try {
+            r3 = getContainingView(r3)
+          } catch (e) {}
+          const featureWidget = session.addWidget(
+            'SyntenyFeatureWidget',
+            'syntenyFeature',
+            {
+              featureData: feature.toJSON(),
+              view: r3,
+              track: getContainingTrack(self),
+            },
+          )
+          session.showWidget(featureWidget)
+        }
+        session.setSelection(feature)
+      },
       afterCreate() {
         // use color by stand to help indicate inversions better on first load,
         // otherwise use selected orientation
