+++ plugins/linear-comparative-view/src/LinearSyntenyDisplay/components/LinearSyntenyRendering.tsx
@@ -200,7 +200,13 @@ const LinearSyntenyRendering = observer(function ({
               const { f, cigar } = model.featPositions[id]
               const unitMultiplier2 = Math.floor(MAX_COLOR_RANGE / cigar.length)
               const cigarIdx = getId(r2!, g2!, b2!, unitMultiplier2)
-              setTooltip(getTooltip(f, cigar[cigarIdx], cigar[cigarIdx + 1]))
+              setTooltip(
+                getTooltip({
+                  feature: f,
+                  cigarOp: cigar[cigarIdx],
+                  cigarOpLen: cigar[cigarIdx + 1],
+                }),
+              )
             }
           }
         }}
