+++ plugins/linear-comparative-view/src/LinearSyntenyDisplay/components/SyntenyContextMenu.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import { getContainingView, getSession, Feature } from '@jbrowse/core/util'
 import { Menu } from '@jbrowse/core/ui'
-import { getParent } from 'mobx-state-tree'
 
 // locals
 import { LinearSyntenyDisplayModel } from '../model'
@@ -55,14 +54,13 @@ export default function SyntenyContextMenu({
           label: 'Center on feature',
           onClick: () => {
             const { f } = feature
-            const track = getParent<{ level: number }>(model, 4)
             const start = f.get('start')
             const end = f.get('end')
             const refName = f.get('refName')
             const mate = f.get('mate')
 
-            const l1 = view.views[track.level]!
-            const l2 = view.views[track.level + 1]!
+            const l1 = view.views[model.level]!
+            const l2 = view.views[model.level + 1]!
             l1.navToLocString(`${refName}:${start}-${end}`).catch(
               (e: unknown) => {
                 const err = `${l1.assemblyNames[0]}:${e}`
